% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice-pc-set.R
\name{all_pc_set_voicings}
\alias{all_pc_set_voicings}
\title{Pitch-class set voicings}
\usage{
all_pc_set_voicings(pc_set, min_octave, max_octave)
}
\arguments{
\item{pc_set}{Numeric vector; the pitch-class set to voice.}

\item{min_octave}{Integerish scalar;
the minimum octave allowed in the voicing,
expressed relative to middle C (MIDI note 60).
For example, -1 identifies the octave ranging from one octave below middle C
to the B below middle C.}

\item{max_octave}{Integerish scalar;
the maximum octave allowed in the voicing,
expressed relative to middle C (MIDI note 60).
For example, 0 identifies the octave ranging from middle C
to the B 11 semitones above.}
}
\value{
A list of all possible voicings, expressed as numeric vectors
of MIDI note numbers.
}
\description{
Finds all possible voicings of a pitch-class set within a given octave range.
}
