% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set.R
\name{all_voicings_pc_set}
\alias{all_voicings_pc_set}
\title{All voicings (pitch-class set)}
\usage{
all_voicings_pc_set(x, opt)
}
\arguments{
\item{x}{Pitch-class set to voice.}

\item{opt}{Options list as created by \code{\link{voice_opt}}.}
}
\value{
A list of possible voicings.
}
\description{
Lists all the possible voicings for an object of class
\code{\link[hrep]{pc_set}}.
}
