% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-change-num-notes.R
\name{change_num_notes}
\alias{change_num_notes}
\title{Change in number of notes}
\usage{
change_num_notes(contexts, continuation)
}
\arguments{
\item{contexts}{A list of chords corresponding to the first chord
in the transition. 
Each chord should be represented as a numeric vector
of MIDI note numbers, without duplicates,
sorted in ascending order.
The function is vectorized over this argument.}

\item{continuation}{The second chord in the transition,
represented as a numeric vector
of MIDI note numbers, without duplicates,
sorted in ascending order.}
}
\value{
An integer vector, vectorized over \code{contexts},
providing the absolute change in the number of chord notes
between the ith element of \code{contexts}
and \code{continuation}.
}
\description{
Checks for changes in the number of notes 
within a chord transition.
}
\note{
This is a \code{voicer} feature suitable for passing
to the \code{features} argument of \code{\link{voice_opt}}.
}
