% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-diff-num-notes.R
\name{diff_num_notes}
\alias{diff_num_notes}
\title{Difference in the number of notes}
\usage{
diff_num_notes(ideal_num_notes)
}
\arguments{
\item{ideal_num_notes}{(Numeric scalar)
The ideal number of notes in the chord.}
}
\value{
A numeric scalar corresponding to the difference between
the number of notes in the chord and
an ideal number of notes as determined by the
\code{ideal_num_notes} argument.
}
\description{
Defines and returns a \code{voicer} feature, suitable for passing
to the \code{features} argument of \code{\link{voice_opt}},
that computes how much the number of notes in a chord voicing
differs from an ideal number of notes as specified by the
\code{ideal_num_notes} argument.
The resulting function takes one argument, \code{x},
corresponding to a chord represented as a numeric vector
of MIDI note numbers, without duplicates,
sorted in ascending order.
}
\note{
The resulting function is not vectorized.
}
