% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-dist.R
\name{dist_below_bottom}
\alias{dist_below_bottom}
\title{Distance below bottom}
\usage{
dist_below_bottom(bottom)
}
\arguments{
\item{bottom}{(Numeric scalar)
The pitch height to which the chord is compared,
specified as a MIDI note number where 60 corresponds to middle C.}
}
\description{
Defines and returns a \code{voicer} feature, suitable for passing
to the \code{features} argument of \code{\link{voice_opt}},
that computes the extent to which a chord spans 
below a predefined pitch height specified in the 
\code{bottom} argument, in semitones.
If the chord does not span below \code{bottom},
then the function returns \code{0}.
The resulting function takes one argument, \code{x},
corresponding to a chord represented as a numeric vector
of MIDI note numbers, without duplicates,
sorted in ascending order.
}
\note{
The resulting function is not vectorized.
}
