% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-consonance.R
\name{hutch_78}
\alias{hutch_78}
\title{Dissonance}
\usage{
hutch_78(x)
}
\arguments{
\item{x}{The chord to analyze,
expressed as a numeric vector of MIDI note numbers
in ascending order.}
}
\value{
A numeric scalar dissonance estimate.
}
\description{
Estimates the dissonance of a chord voicing using 
Hutchinson and Knopoff's (1978) algorithm as imported 
from the \code{\link{dycon}} package.
The function is memoized to make repeated calls more efficient.
}
\note{
This is a \code{voicer} feature suitable for passing
to the \code{features} argument of \code{\link{voice_opt}}.
}
