% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-corpus-features.R
\name{is_corpus_features}
\alias{is_corpus_features}
\title{Is it a corpus features object?}
\usage{
is_corpus_features(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of class \code{corpus_features},
as produced by \code{\link{get_corpus_features}}.
}
