% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-features.R
\name{is_voicer_model}
\alias{is_voicer_model}
\title{Is it a voicer model object?}
\usage{
is_voicer_model(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of class \code{voicer_model},
as produced by \code{\link{model_features}}.
}
