% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-features.R
\name{is_voicer_weights}
\alias{is_voicer_weights}
\title{Is it a voicer weights object?}
\usage{
is_voicer_weights(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Checks whether an object is of class \code{voicer_weights},
as found in the \code{weights} slot of a \code{voicer_model}
object produced by \code{\link{model_features}}.
}
