% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-features.R
\name{linear_formula}
\alias{linear_formula}
\title{Linear formula}
\usage{
linear_formula(x, features)
}
\arguments{
\item{x}{(Data frame)
Dataset with respect to which features are defined.}

\item{features}{Character vector of features,
with each feature corresponding to a unique column 
of \code{x}.}
}
\value{
A formula object to be passed to \code{\link{model_features}}.
}
\description{
Defines a formula for the \code{formula} argument of 
\code{\link{model_features}} corresponding
to a simple linear combination of features.
}
