% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalityPlots.R
\name{normalityPlots}
\alias{normalityPlots}
\title{Normality Plots}
\usage{
normalityPlots(
  audioData,
  measures = c("duration", "voice_breaks_percent", "RMS_env", "mean_loudness", "mean_F0",
    "sd_F0", "mean_entropy", "mean_HNR")
)
}
\arguments{
\item{audioData}{A data.frame generated by autoExtract.}

\item{measures}{An optional character vector indicating the name of the variables to be plotted.}
}
\value{
A list containing the different plots that are generated.
}
\description{
Generates plots showing the normality of the different measures from the data.frame obtained from autoExtract.
}
\examples{
normalityPlots(testAudioData)

}
