% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-parallels.R
\name{outer_parallels}
\alias{outer_parallels}
\title{Outer parallels}
\usage{
outer_parallels(contexts, continuation)
}
\arguments{
\item{contexts}{See \code{\link{any_parallels}}.}

\item{continuation}{See \code{\link{any_parallels}}.}
}
\description{
Equivalent to \code{\link{any_parallels}},
but only returns \code{TRUE} when the parallel octaves
are found between the bass and melody lines.
}
