% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess list of Audio objects}
\usage{
preprocess(audioList, normalizeAmplitude = TRUE, removeNoise = TRUE, ...)
}
\arguments{
\item{audioList}{A list of Wave objects.}

\item{normalizeAmplitude}{A logical value indicating whether to normalize amplitude.}

\item{removeNoise}{A logical value indicating whether to remove background noise.}

\item{...}{Other options used to control preprocessing behavior.}
}
\value{
A list of (processed) Wave objects.
}
\description{
Automatically preprocesses a list of Wave objects by normalizing their amplitude and removing background noise.
}
\examples{
preprocess(testAudioList)

}
