% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-features.R
\name{profile_features}
\alias{profile_features}
\title{Profile features}
\usage{
profile_features(random = TRUE, features = voice_features())
}
\arguments{
\item{random}{(Logical scalar)
Whether the two chords should be resampled each time or not.
The latter is useful for evaluating the effect of memoization on function speeds.}

\item{features}{A named list of feature functions, defaulting to the list created by 
\code{\link{voice_features}}.}
}
\value{
A named numeric vector of computation times for each feature,
with units of seconds.
}
\description{
Computes profiling information for a set of features.
These features are computed for a transition between two chords.
}
