% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAudio.R
\name{saveAudio}
\alias{saveAudio}
\title{Save Audio Files}
\usage{
saveAudio(audioList, path = "./Preprocessed/", fileType = "wav")
}
\arguments{
\item{audioList}{A list of Wave objects.}

\item{path}{A character string indicating the full path to the folder containing the audio files. Default corresponds to the current working directory.}

\item{fileType}{Character string indicating the file format (wav or mp3) of the audio files. Default corresponds to wav.}
}
\value{
Save objects of a \code{Wave} list as .mp3 or .wav files.
}
\description{
Save all objects in a list of \code{Wave} objects as a .wav or .mp3 file in the specified path.
}
\examples{
\donttest{
saveAudio(testAudioList, fileType = "wav")
}
}
