% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableANOVA.R
\name{tableANOVA}
\alias{tableANOVA}
\title{Create a Table of ANOVA results}
\usage{
tableANOVA(
  audioData,
  by = c(),
  measure = "duration",
  nameMeasure = c(),
  figureNumber = 1
)
}
\arguments{
\item{audioData}{A data.frame generated by the autoExtract() function.}

\item{by}{A character vector indicating the name of the factor(s).}

\item{measure}{Name of the dependent variable.}

\item{nameMeasure}{Optional relabelling of dependent variable for the output table. If no value is provided, the original variable name is used.}

\item{figureNumber}{An integer indicating the figure number to create the title for the table. Default corresponds to 1.}
}
\value{
HTML table showing the ANOVA results in APA formatting style.
}
\description{
Automatically generates HTML table with results for one-way or two-way ANOVAs.
}
\examples{
tableANOVA(testAudioData, by = c("Condition", "Dimension"), measure = "duration")

}
