% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-vl-dist.R
\name{vl_dist}
\alias{vl_dist}
\title{Voice-leading distance}
\usage{
vl_dist(contexts, continuation)
}
\arguments{
\item{contexts}{A list of chords corresponding to the first chord
in the transition. 
Each chord should be represented as a numeric vector
of MIDI note numbers, without duplicates,
sorted in ascending order.
The function is vectorized over this argument.}

\item{continuation}{The second chord in the transition,
represented as a numeric vector
of MIDI note numbers, without duplicates,
sorted in ascending order.}
}
\value{
A numeric vector of the same length as \code{contexts},
with the ith element corresponding to the sum voice-leading distance
between the ith element of \code{contexts} and \code{continuation}.
}
\description{
Computes the minimal voice-leading distance (in semitones)
between two chords,
using the minimal voice-leading algorithm
of \insertCite{Tymoczko2006;textual}{voicer}
as implemented in the \code{minVL} package (see \code{\link[minVL]{min_vl}}).
}
\note{
Unlike Tymoczko's original presentation, 
distances are computed in pitch space, not pitch-class space.
}
\references{
\insertAllCited{}
}
