% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice.R
\name{voice}
\alias{voice}
\title{Voice}
\usage{
voice(x, opt = voice_opt())
}
\arguments{
\item{x}{Chord sequence to voice.
This should take the form of a \code{\link[hrep]{vec}}
object from the \code{hrep} package.
A \code{\link[hrep]{vec}} object is essentially a typed list;
here the type should be one of three chord representations from
the \code{hrep} package, namely
\code{\link{pc_chord}}, \code{\link{pc_set}}, or \code{\link{pi_chord}}.
See \bold{Note} for more information.}

\item{opt}{A list of options as created by \code{\link{voice_opt}}.}
}
\value{
A voiced chord sequence,
represented as a \code{\link[hrep]{vec}} object 
with type \code{\link[hrep]{pi_chord}}.
This object can be inspected using \code{\link[base]{as.list}}.
}
\description{
Generates an optimized voicing for a chord sequence.
}
\details{
By default, the algorithm optimizes the sum of the linear predictors
for each chord transition.
By setting the \code{exp_cost}, \code{norm_cost}, and \code{log_cost}
arguments of \code{\link{voice_opt}} to \code{TRUE},
it is possible alternatively to optimize the probability of the sequence.
}
\note{
Voicing sequences of type \code{\link{pi_chord}} preserves 
the original bass pitch classes.
The original doublings for each pitch class can also be preserved
by setting \code{dbl_change = TRUE} in \code{\link{voice_opt}}.
}
\examples{
library(magrittr)
library(hrep)
list(
  pc_chord(c(0, 4, 7)),
  pc_chord(c(0, 5, 9)),
  pc_chord(c(2, 7, 11))
) \%>\%
  vec("pc_chord") \%>\%
  voice(voice_opt(min_notes = 3, max_notes = 3))

}
