% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice-opt.R
\docType{data}
\name{voice_default_weights}
\alias{voice_default_weights}
\title{Default weights}
\format{An object of class \code{numeric} of length 12.}
\usage{
voice_default_weights
}
\description{
Provides the default regression weights for the features 
listed in \code{\link{voice_features}}.
These were optimized on the \code{\link[hcorp]{bach_chorales_1}} dataset.
}
\details{
The object constitutes a named numeric vector,
ordered to match \code{\link{voice_features}},
where the names identify the features and the
values correspond to the regression weights.
These weights are suitable for passing to the \code{weights} argument
of \code{\link{voice_opt}}.
}
\keyword{data}
