% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice-cost-funs.R
\name{voice_features}
\alias{voice_features}
\title{Features}
\usage{
voice_features(top = 72, middle = 60, bottom = 48,
  ideal_num_notes = 4L)
}
\arguments{
\item{top}{(Numeric scalar)
Passed to \code{\link{dist_above_top}}.}

\item{middle}{(Numeric scalar)
Passed to \code{\link{dist_from_middle}}.}

\item{bottom}{(Numeric scalar)
Passed to \code{\link{dist_below_bottom}}.}

\item{ideal_num_notes}{Passed to \code{\link{diff_num_notes}}.}
}
\value{
Returns a named list of features suitable for passing to 
\code{\link{voice_opt}} or \code{\link{get_corpus_features}}.
}
\description{
Defines a default set of features for use by 
\code{\link{voice}} and \code{\link{get_corpus_features}}.
}
