% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice-opt.R
\name{voice_opt}
\alias{voice_opt}
\title{Voicing options}
\usage{
voice_opt(min_octave = -2L, max_octave = 1L, dbl_change = TRUE,
  min_notes = 1L, max_notes = 4L, features = voice_features(),
  weights = voice_default_weights, exp_cost = FALSE,
  norm_cost = FALSE, log_cost = FALSE, verbose = interactive())
}
\arguments{
\item{min_octave}{(Integerish scalar)
The minimum octave allowed in the voicing,
expressed relative to middle C (MIDI note 60).
For example, -1 identifies the octave ranging from one octave below middle C
to the B below middle C.}

\item{max_octave}{(Integerish scalar)
The maximum octave allowed in the voicing,
expressed relative to middle C (MIDI note 60).
For example, 0 identifies the octave ranging from middle C
to the B 11 semitones above.}

\item{dbl_change}{(Logical scalar)
Is it permitted to change the doubling of the chords,
whether by adding duplicated pitch classes or removing duplicated pitch classes?}

\item{min_notes}{(Integerish scalar)
Sets the minimum allowed number of notes in the voiced chords;
ignored if \code{dbl_change = FALSE}.}

\item{max_notes}{(Integerish scalar)
Sets the maximum allowed number of notes in the voiced chords;
ignored if \code{dbl_change = FALSE}.}

\item{features}{A named list of features to apply to chord transitions,
defaulting to a list created by \code{\link{voice_features}}.
Any new features must be created using \code{\link[seqopt]{cost_fun}},
and take chords as arguments,
where the chords are represented as ordered numeric vectors 
of MIDI note numbers.}

\item{weights}{A numeric vector of weights mapping to the respective elements
of \code{features} in their original order.
This vector need not be named,
but an error will be thrown if names are provided
and they do not match with those of \code{features}.}

\item{exp_cost}{(Logical scalar)
Experimental - passed to \code{\link[seqopt]{seq_opt}},
thereby determining whether the linear predictor is exponentiated
when computing a sequence's cost.}

\item{norm_cost}{(Logical scalar)
Experimental - passed to \code{\link[seqopt]{seq_opt}},
thereby determining whether the (possibly exponentiated) linear predictor 
is normalized over all possible continuations
when computing a sequence's cost.}

\item{log_cost}{(Logical scalar)
Experimental - passed to \code{\link[seqopt]{seq_opt}},
thereby determining whether to take the logarithm of the
(possibly exponentiated, possibly normalized) linear predictor 
when computing a sequence's cost.}

\item{verbose}{(Logical scalar)
Determines whether progress messages are printed during the 
function's execution.}
}
\description{
Defines a list of options to be passed to \code{\link{voice}}.
}
