% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voicer-package.R
\docType{package}
\name{voicer-package}
\alias{voicer}
\alias{voicer-package}
\title{voicer: Analyzing and Generating Chord Voicings}
\description{
This package supports the analysis and generation of voicings for chord sequences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pmcharrison/voicer}
  \item Report bugs at \url{https://github.com/pmcharrison/voicer/issues}
}

}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
