% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceRApp.R
\name{voiceRApp}
\alias{voiceRApp}
\title{Start voiceR Shiny App}
\usage{
voiceRApp()
}
\value{
This function launches the voiceR Shiny app.
}
\description{
Launches the voiceR Shiny app, providing the opportunity to analyze multiple audio files via a graphical user interface (no-code interface).
}
\examples{
if(interactive()){
   voiceRApp()
}


}
