% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set.R
\name{all_voicings_pc_multiset}
\alias{all_voicings_pc_multiset}
\title{All voicings (pitch-class multiset)}
\usage{
all_voicings_pc_multiset(x, opt)
}
\arguments{
\item{x}{Pitch-class multiset to voice, expressed as a numeric vector
with potentially repeated elements.}

\item{opt}{Options list as created by \code{\link{voice_opt}}.}
}
\value{
A list of possible voicings.
}
\description{
Lists all the possible voicings for a pitch-class multiset.
}
