% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice-pc-set.R
\name{all_voicings_pc_multiset}
\alias{all_voicings_pc_multiset}
\title{All voicings (pitch-class multiset)}
\usage{
all_voicings_pc_multiset(x, min_octave, max_octave)
}
\arguments{
\item{x}{Pitch-class multiset to voice, expressed as a numeric vector
with potentially repeated elements.}
}
\value{
A list of possible voicings.
}
\description{
Lists all the possible voicings for a pitch-class multiset.
}
