% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-set.R
\name{all_voicings_pc_set}
\alias{all_voicings_pc_set}
\title{All voicings (pitch-class set)}
\usage{
all_voicings_pc_set(x, min_octave, max_octave, dbl_change, dbl_min,
  dbl_max)
}
\arguments{
\item{x}{Pitch-class set to voice.}
}
\value{
A list of possible voicings.
}
\description{
Lists all the possible voicings for an object of class
\code{\link[hrep]{pc_set}}.
}
