% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord.R
\name{all_voicings_pi_chord}
\alias{all_voicings_pi_chord}
\title{All voicings}
\usage{
all_voicings_pi_chord(x, opt)
}
\arguments{
\item{x}{Object to voice.}

\item{opt}{Options list as created by \code{\link{voice_opt}}.}
}
\value{
A list of possible voicings.
}
\description{
Lists all the possible voicings for an object of class
\code{\link[hrep]{pi_chord}}.
}
