% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord.R
\name{all_voicings_pi_chord}
\alias{all_voicings_pi_chord}
\title{All voicings}
\usage{
all_voicings_pi_chord(x, min_octave, max_octave, dbl_change, dbl_min,
  dbl_max)
}
\arguments{
\item{x}{Object to voice.}
}
\value{
A list of possible voicings.
}
\description{
Lists all the possible voicings for an object of class
\code{\link[hrep]{pi_chord}}.
}
