% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice-pi-chord.R
\name{all_voicings_pi_chord}
\alias{all_voicings_pi_chord}
\title{All voicings (pi_chord)}
\usage{
all_voicings_pi_chord(x, min_octave, max_octave, dbl_change, min_notes,
  max_notes, fix_melody = NA_integer_, fix_content = integer(),
  fix_chord = NULL)
}
\arguments{
\item{x}{Object to voice.}

\item{min_octave}{See \code{\link{voice_opt}}.}

\item{max_octave}{See \code{\link{voice_opt}}.}

\item{dbl_change}{See \code{\link{voice_opt}}.}

\item{min_notes}{See \code{\link{voice_opt}}.}

\item{max_notes}{See \code{\link{voice_opt}}.}

\item{fix_melody}{(Numeric scalar)
Determines the MIDI pitch for the melody (i.e. the top note of the voicing).
If NA, no constraint is applied.}

\item{fix_content}{(Numeric vector)
Specifies a set of MIDI pitches that must be contained in the voicing.}

\item{fix_chord}{(NULL or numeric vector)
If not NULL, the function returns just one voicing with the MIDI pitches
specified in this vector.}
}
\value{
A list of possible voicings.
}
\description{
Lists all the possible voicings for an object of class
\code{\link[hrep]{pi_chord}}.
}
\details{
By default, these voicings preserve the chord's bass pitch class.
To include all possible inversions,
transform the chord first to an object of class
\code{\link[hrep]{pc_set}} and then call
\code{\link{all_voicings_pc_set}}.
By setting \code{dbl_change} to \code{TRUE},
it is possible to preserve the pitch-class doublings
in the original chord.
}
